+++ gradle/build-logic/src/main/kotlin/CompileOptions.kt
@@ -11,8 +11,8 @@ object CompileOptions {
 
     object AocKt {
         const val GROUP_ID: String = "io.github.jadarma.aockt"
-        const val CURRENT: String = "0.2.1" // Last released version.
-        const val NEXT: String = "0.3.0" // Current snapshot.
+        const val CURRENT: String = "0.3.0" // Last released version.
+        const val NEXT: String = "0.4.0" // Current snapshot.
     }
 
     object Java {