+++ docs/topics/changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.3.0-SNAPSHOT
 
+- Updated minimum JDK to 21.
 - Updated Kotlin to `2.2.10` and Kotest to `6.0.3`.
 - There are no breaking changes to AocKt itself, but there are for configuring 
   [Kotest 6.0](https://kotest.io/docs/release6).